requirements.txt
openai-whisper
moviepy
streamlit
torch
numpy
ffmpeg-python